refactor: extract pricing and LiteLLM authority
